requirements.txt
bottle==0.13.3
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.9
litecli==1.12.3
prompt_toolkit==3.0.48
Pygments==2.18.0
PyJWT==2.10.1
six==1.16.0
sqlparse==0.5.1
tabulate==0.9.0
wcwidth==0.2.13
